move methods in network.__init__ to network.connector, add __console_dir__
